Se arreglo un error en el ms inventario
error con el id que estaba en integer aun y no long

+++ ms-inventario/src/main/java/cl/hilton/inventario/repository/MinibarRepository.java
@@ -1,6 +1,5 @@
 package cl.hilton.inventario.repository;
 
-
 import cl.hilton.inventario.model.Minibar;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MinibarRepository extends JpaRepository<Minibar, Integer> {
+public interface MinibarRepository extends JpaRepository<Minibar, Long> {
 
     List<Minibar> findByHabitacionNumeroHabitacion(String numeroHabitacion);
